+++ src/sw.js
@@ -1,4 +1,5 @@
 import { skipWaiting, clientsClaim } from 'workbox-core'
+import { precacheAndRoute } from 'workbox-precaching'
 import { registerRoute } from 'workbox-routing'
 import { StaleWhileRevalidate, CacheFirst, NetworkFirst } from 'workbox-strategies'
 import { CacheableResponsePlugin } from 'workbox-cacheable-response'
@@ -6,6 +7,10 @@ import { ExpirationPlugin } from 'workbox-expiration'
 
 skipWaiting()
 clientsClaim()
+
+/* global self */
+
+precacheAndRoute(self.__WB_MANIFEST)
 
 /**
  * Runtime caching